Refactor jules-queue.js into submodules
- Split jules-queue.js into jules-queue-data.js (CRUD), jules-queue-modal.js (UI), and jules-queue-actions.js (logic).
- Retained jules-queue.js as a facade for backward compatibility.
- Updated consumers to import from specific modules where appropriate.
- Verified with tests.
