+++ src/modules/jules-account.js
@@ -3,7 +3,7 @@
 
 import { checkJulesKey, deleteStoredJulesKey } from './jules-keys.js';
 import { showJulesKeyModal } from './jules-modal.js';
-import { showJulesQueueModal } from './jules-queue.js';
+import { showJulesQueueModal } from './jules-queue-modal.js';
 import { loadJulesProfileInfo, listJulesSessions, getDecryptedJulesKey } from './jules-api.js';
 import { getCache, setCache, clearCache, CACHE_KEYS } from '../utils/session-cache.js';
 import { showToast } from './toast.js';
