+++ src/modules/jules-queue-actions.js
@@ -1,0 +1,508 @@
+import { extractTitleFromPrompt } from '../utils/title.js';
+import statusBar from './status-bar.js';
+import { showToast } from './toast.js';
+import { handleError, ErrorCategory } from '../utils/error-handler.js';
+import { showConfirm } from './confirm-modal.js';
+import { JULES_MESSAGES, TIMEOUTS } from '../utils/constants.js';
+import { callRunJulesFunction } from './jules-api.js';
+import { openUrlInBackground, showSubtaskErrorModal } from './jules-modal.js';
+import { getServerTimestamp, getFieldDelete } from '../utils/firestore-helpers.js';
+import { CACHE_KEYS, clearCache } from '../utils/session-cache.js';
+
+import {
+  addToJulesQueue,
+  updateJulesQueueItem,
+  deleteFromJulesQueue,
+  getQueueCache
+} from './jules-queue-data.js';
+
+import {
+  loadQueuePage,
+  getSelectedQueueIds
+} from './jules-queue-modal.js';
+
+export async function handleQueueAction(queueItemData) {
+  const user = window.auth?.currentUser;
+  if (!user) {
+    handleError(JULES_MESSAGES.SIGN_IN_REQUIRED, { source: 'handleQueueAction' }, { category: ErrorCategory.AUTH, toastType: 'warn' });
+    return false;
+  }
+  try {
+    await addToJulesQueue(user.uid, queueItemData);
+    showToast(JULES_MESSAGES.QUEUED, 'success');
+    return true;
+  } catch (err) {
+    handleError(err, { source: 'handleQueueAction' });
+    return false;
+  }
+}
+
+export async function deleteSelectedSubtasks(docId, indices) {
+  const user = window.auth?.currentUser;
+  if (!user) return;
+
+  const queueCache = getQueueCache();
+  const item = queueCache.find(i => i.id === docId);
+  if (!item || !Array.isArray(item.remaining)) return;
+
+  const sortedIndices = indices.sort((a, b) => b - a);
+  const newRemaining = item.remaining.slice();
+
+  for (const index of sortedIndices) {
+    if (index >= 0 && index < newRemaining.length) {
+      newRemaining.splice(index, 1);
+    }
+  }
+
+  if (newRemaining.length === 0) {
+    await deleteFromJulesQueue(user.uid, docId);
+  } else {
+    await updateJulesQueueItem(user.uid, docId, {
+      remaining: newRemaining,
+      updatedAt: getServerTimestamp()
+    });
+  }
+}
+
+export async function runSelectedSubtasks(docId, indices, suppressPopups = false, openInBackground = false) {
+  const queueCache = getQueueCache();
+  const item = queueCache.find(i => i.id === docId);
+  if (!item || !Array.isArray(item.remaining)) return;
+
+  const sortedIndices = indices.sort((a, b) => a - b);
+  const toRun = sortedIndices.map(i => item.remaining[i]).filter(Boolean);
+
+  const successfulIndices = [];
+  const skippedIndices = [];
+
+  for (let i = 0; i < toRun.length; i++) {
+    const subtask = toRun[i];
+    const originalIndex = sortedIndices[i];
+    let retry = true;
+
+    while (retry) {
+      try {
+        const title = extractTitleFromPrompt(subtask.fullContent);
+        const sessionUrl = await callRunJulesFunction(subtask.fullContent, item.sourceId, item.branch || 'master', title);
+        if (sessionUrl && !suppressPopups && item.autoOpen !== false) {
+          if (openInBackground) {
+            openUrlInBackground(sessionUrl);
+          } else {
+            window.open(sessionUrl, '_blank', 'noopener,noreferrer');
+          }
+        }
+        successfulIndices.push(originalIndex);
+        await new Promise(r => setTimeout(r, TIMEOUTS.queueDelay));
+        retry = false;
+      } catch (err) {
+        const result = await showSubtaskErrorModal(i + 1, toRun.length, err, true);
+
+        if (result.action === 'retry') {
+          if (result.shouldDelay) await new Promise(r => setTimeout(r, TIMEOUTS.longDelay));
+          continue;
+        } else if (result.action === 'skip') {
+          skippedIndices.push(originalIndex);
+          retry = false;
+        } else if (result.action === 'queue') {
+          if (successfulIndices.length > 0) {
+            await deleteSelectedSubtasks(docId, successfulIndices);
+          }
+          return { successful: successfulIndices.length, skipped: skippedIndices.length };
+        } else {
+          if (successfulIndices.length > 0) {
+            await deleteSelectedSubtasks(docId, successfulIndices);
+          }
+          const err = new Error('User cancelled');
+          err.successfulCount = successfulIndices.length;
+          throw err;
+        }
+      }
+    }
+  }
+
+  if (successfulIndices.length > 0) {
+    await deleteSelectedSubtasks(docId, successfulIndices);
+  }
+
+  return { successful: successfulIndices.length, skipped: skippedIndices.length };
+}
+
+export async function unscheduleSelectedQueueItems() {
+  const user = window.auth?.currentUser;
+  if (!user) {
+    handleError(JULES_MESSAGES.NOT_SIGNED_IN, { source: 'unscheduleSelectedQueueItems' }, { category: ErrorCategory.AUTH });
+    return;
+  }
+
+  const { queueSelections } = getSelectedQueueIds();
+
+  if (queueSelections.length === 0) {
+    showToast('No items selected', 'warn');
+    return;
+  }
+
+  const itemText = queueSelections.length === 1 ? 'item' : 'items';
+  const confirmed = await showConfirm(`Unschedule ${queueSelections.length} selected ${itemText}?`, {
+    title: 'Unschedule Items',
+    confirmText: 'Unschedule',
+    confirmStyle: 'warn'
+  });
+
+  if (!confirmed) return;
+
+  try {
+    const batch = firebase.firestore().batch();
+
+    for (const docId of queueSelections) {
+      const docRef = firebase.firestore().collection('julesQueues').doc(user.uid).collection('items').doc(docId);
+      batch.update(docRef, {
+        status: 'pending',
+        scheduledAt: getFieldDelete(),
+        scheduledTimeZone: getFieldDelete(),
+        activatedAt: getFieldDelete(),
+        updatedAt: getServerTimestamp()
+      });
+    }
+
+    await batch.commit();
+
+    clearCache(CACHE_KEYS.QUEUE_ITEMS, user.uid);
+
+    showToast(`${queueSelections.length} ${itemText} unscheduled`, 'success');
+    await loadQueuePage();
+  } catch (err) {
+    handleError(err, { source: 'unscheduleSelectedQueueItems' });
+  }
+}
+
+export async function deleteSelectedQueueItems() {
+  const user = window.auth?.currentUser;
+  if (!user) { handleError(JULES_MESSAGES.NOT_SIGNED_IN, { source: 'deleteSelectedQueueItems' }, { category: ErrorCategory.AUTH }); return; }
+
+  const { queueSelections, subtaskSelections } = getSelectedQueueIds();
+
+  if (queueSelections.length === 0 && Object.keys(subtaskSelections).length === 0) {
+    showToast('No items selected', 'warn');
+    return;
+  }
+
+  const totalCount = queueSelections.length + Object.values(subtaskSelections).reduce((sum, arr) => sum + arr.length, 0);
+  const confirmed = await showConfirm(`Delete ${totalCount} selected item(s)?`, {
+    title: 'Delete Items',
+    confirmText: 'Delete',
+    confirmStyle: 'error'
+  });
+  if (!confirmed) return;
+
+  try {
+    for (const id of queueSelections) {
+      await deleteFromJulesQueue(user.uid, id);
+    }
+
+    for (const [docId, indices] of Object.entries(subtaskSelections)) {
+      if (queueSelections.includes(docId)) continue;
+
+      await deleteSelectedSubtasks(docId, indices);
+    }
+
+    showToast(JULES_MESSAGES.deleted(totalCount), 'success');
+    await loadQueuePage();
+  } catch (err) {
+    handleError(err, { source: 'deleteSelectedQueueItems' });
+  }
+}
+
+function sortByCreatedAt(ids) {
+  const queueCache = getQueueCache();
+  return ids.slice().sort((a, b) => {
+    const itemA = queueCache.find(i => i.id === a);
+    const itemB = queueCache.find(i => i.id === b);
+    return (itemA?.createdAt?.seconds || 0) - (itemB?.createdAt?.seconds || 0);
+  });
+}
+
+export async function runSelectedQueueItems() {
+  const user = window.auth?.currentUser;
+  if (!user) { handleError(JULES_MESSAGES.NOT_SIGNED_IN, { source: 'runSelectedQueueItems' }, { category: ErrorCategory.AUTH }); return; }
+
+  const { queueSelections, subtaskSelections } = getSelectedQueueIds();
+
+  if (queueSelections.length === 0 && Object.keys(subtaskSelections).length === 0) {
+    showToast('No items selected', 'warn');
+    return;
+  }
+
+  const suppressPopups = document.getElementById('queueSuppressPopupsCheckbox')?.checked || false;
+  const openInBackground = document.getElementById('queueOpenInBackgroundCheckbox')?.checked || false;
+  const pauseBtn = document.getElementById('queuePauseBtn');
+  let paused = false;
+  if (pauseBtn) {
+    pauseBtn.disabled = false;
+    pauseBtn.onclick = () => {
+      paused = true;
+      pauseBtn.disabled = true;
+      statusBar.showMessage('Pausing queue processing after the current subtask', { timeout: TIMEOUTS.longDelay });
+    };
+  }
+
+  statusBar.showMessage('Processing queue...', { timeout: 0 });
+  statusBar.setAction('Pause', () => {
+    paused = true;
+    statusBar.showMessage('Pausing after current subtask', { timeout: TIMEOUTS.statusBar });
+    statusBar.clearAction();
+    if (pauseBtn) pauseBtn.disabled = true;
+  });
+
+  const queueCache = getQueueCache();
+  const sortedSubtaskEntries = Object.entries(subtaskSelections).sort(([a], [b]) =>
+    (queueCache.find(i => i.id === a)?.createdAt?.seconds || 0) - (queueCache.find(i => i.id === b)?.createdAt?.seconds || 0)
+  );
+  const totalSubtasks = Object.entries(subtaskSelections)
+    .filter(([docId]) => !queueSelections.includes(docId))
+    .reduce((sum, [, indices]) => sum + indices.length, 0);
+  const totalSingles = queueSelections.filter(id => {
+    const item = queueCache.find(i => i.id === id);
+    return item?.type === 'single';
+  }).length;
+  const totalFullSubtaskQueues = queueSelections.filter(id => {
+    const item = queueCache.find(i => i.id === id);
+    return item?.type === 'subtasks';
+  }).reduce((sum, id) => {
+    const item = queueCache.find(i => i.id === id);
+    return sum + (item?.remaining?.length || 0);
+  }, 0);
+  const totalItems = totalSubtasks + totalSingles + totalFullSubtaskQueues;
+  let currentItemNumber = 0;
+  let totalSkipped = 0;
+  let totalSuccessful = 0;
+
+  for (const [docId, indices] of sortedSubtaskEntries) {
+    if (paused) break;
+    if (queueSelections.includes(docId)) continue;
+
+    // const item = queueCache.find(i => i.id === docId); // Unused
+
+    try {
+      const result = await runSelectedSubtasks(docId, indices.slice().sort((a, b) => a - b), suppressPopups, openInBackground);
+      currentItemNumber += indices.length;
+      if (result && result.skipped > 0) {
+        totalSkipped += result.skipped;
+      }
+      if (result && result.successful > 0) {
+        totalSuccessful += result.successful;
+      }
+    } catch (err) {
+      if (err.message === 'User cancelled') {
+        if (err.successfulCount) {
+          totalSuccessful += err.successfulCount;
+        }
+        showToast(JULES_MESSAGES.cancelled(totalSuccessful, totalItems), 'warn');
+        statusBar.clear();
+        await loadQueuePage();
+        return;
+      }
+      throw err;
+    }
+  }
+
+  for (const id of sortByCreatedAt(queueSelections)) {
+    if (paused) break;
+    const item = queueCache.find(i => i.id === id);
+    if (!item) continue;
+
+    try {
+      if (item.type === 'single') {
+        currentItemNumber++;
+        let retry = true;
+        while (retry) {
+          try {
+            const title = extractTitleFromPrompt(item.prompt || '');
+            const sessionUrl = await callRunJulesFunction(item.prompt || '', item.sourceId, item.branch || 'master', title);
+            if (sessionUrl && !suppressPopups && item.autoOpen !== false) {
+              if (openInBackground) {
+                openUrlInBackground(sessionUrl);
+              } else {
+                window.open(sessionUrl, '_blank', 'noopener,noreferrer');
+              }
+            }
+            await deleteFromJulesQueue(user.uid, id);
+            totalSuccessful++;
+            retry = false;
+          } catch (singleErr) {
+            const result = await showSubtaskErrorModal(currentItemNumber, totalItems, singleErr, true);
+            if (result.action === 'retry') {
+              if (result.shouldDelay) await new Promise(r => setTimeout(r, TIMEOUTS.longDelay));
+              continue;
+            } else if (result.action === 'skip') {
+              totalSkipped++;
+              retry = false;
+            } else if (result.action === 'queue') {
+              retry = false;
+            } else {
+              showToast(JULES_MESSAGES.cancelled(totalSuccessful, totalItems), 'warn');
+              statusBar.clear();
+              await loadQueuePage();
+              return;
+            }
+          }
+        }
+      } else if (item.type === 'subtasks') {
+        let remaining = Array.isArray(item.remaining) ? item.remaining.slice() : [];
+        const skippedSubtasks = [];
+
+        const initialCount = remaining.length;
+        let subtaskNumber = 0;
+        while (remaining.length > 0) {
+          if (paused) {
+            try {
+              await updateJulesQueueItem(user.uid, id, {
+                remaining,
+                status: 'paused',
+                updatedAt: getServerTimestamp()
+              });
+            } catch (e) {
+              console.warn('Failed to persist paused state for queue item', id, e.message || e);
+            }
+            statusBar.showMessage('Paused — progress saved', { timeout: TIMEOUTS.statusBar });
+            statusBar.clearProgress();
+            statusBar.clearAction();
+            await loadQueuePage();
+            return;
+          }
+
+          currentItemNumber++;
+          subtaskNumber++;
+          const s = remaining[0];
+          let subtaskRetry = true;
+          while (subtaskRetry) {
+            try {
+              const title = extractTitleFromPrompt(s.fullContent);
+              const sessionUrl = await callRunJulesFunction(s.fullContent, item.sourceId, item.branch || 'master', title);
+              if (sessionUrl && !suppressPopups && item.autoOpen !== false) {
+                if (openInBackground) {
+                  openUrlInBackground(sessionUrl);
+                } else {
+                  window.open(sessionUrl, '_blank', 'noopener,noreferrer');
+                }
+              }
+
+              remaining.shift();
+              totalSuccessful++;
+
+              try {
+                await updateJulesQueueItem(user.uid, id, {
+                  remaining,
+                  status: remaining.length === 0 ? 'done' : 'in-progress',
+                  updatedAt: getServerTimestamp()
+                });
+              } catch (e) {
+                console.warn('Failed to persist progress for queue item', id, e.message || e);
+              }
+
+              try {
+                const done = initialCount - remaining.length;
+                const percent = initialCount > 0 ? Math.round((done / initialCount) * 100) : 100;
+                statusBar.setProgress(`${done}/${initialCount}`, percent);
+                statusBar.showMessage(`Processing subtask ${done}/${initialCount}`, { timeout: 0 });
+              } catch (e) {
+                console.error('Failed to update UI progress for queue item', id, e);
+              }
+
+              await new Promise(r => setTimeout(r, TIMEOUTS.queueDelay));
+              subtaskRetry = false;
+            } catch (err) {
+              const result = await showSubtaskErrorModal(subtaskNumber, initialCount, err, true);
+
+              if (result.action === 'retry') {
+                if (result.shouldDelay) await new Promise(r => setTimeout(r, TIMEOUTS.longDelay));
+                continue;
+              } else if (result.action === 'skip') {
+                totalSkipped++;
+                skippedSubtasks.push(remaining.shift());
+                try {
+                  await updateJulesQueueItem(user.uid, id, {
+                    remaining: [...skippedSubtasks, ...remaining],
+                    status: 'pending',
+                    updatedAt: getServerTimestamp()
+                  });
+                } catch (e) {
+                  console.warn('Failed to persist remaining after skip', e);
+                }
+                statusBar.showMessage(JULES_MESSAGES.SKIPPED_SUBTASK, { timeout: TIMEOUTS.actionFeedback });
+                subtaskRetry = false;
+              } else if (result.action === 'queue') {
+                try {
+                  await updateJulesQueueItem(user.uid, id, {
+                    remaining,
+                    status: 'pending',
+                    updatedAt: getServerTimestamp()
+                  });
+                } catch (e) {
+                  console.warn('Failed to persist queue state', e);
+                }
+                statusBar.showMessage('Remainder queued for later', { timeout: TIMEOUTS.statusBar });
+                statusBar.clearProgress();
+                statusBar.clearAction();
+                await loadQueuePage();
+                return;
+              } else {
+                try {
+                  await updateJulesQueueItem(user.uid, id, {
+                    remaining,
+                    status: 'error',
+                    updatedAt: getServerTimestamp()
+                  });
+                } catch (e) {
+                  console.warn('Failed to persist error state', e);
+                }
+                showToast(JULES_MESSAGES.cancelled(totalSuccessful, totalItems), 'warn');
+                statusBar.clear();
+                await loadQueuePage();
+                return;
+              }
+            }
+          }
+        }
+
+        if (skippedSubtasks.length > 0 && remaining.length === 0) {
+          try {
+            await updateJulesQueueItem(user.uid, id, {
+              remaining: skippedSubtasks,
+              status: 'pending',
+              updatedAt: getServerTimestamp()
+            });
+          } catch (e) {
+            console.warn('Failed to save skipped subtasks', e);
+          }
+        } else if (skippedSubtasks.length === 0 && remaining.length === 0) {
+          await deleteFromJulesQueue(user.uid, id);
+        }
+      } else {
+        console.warn('Unknown queue item type', item.type);
+      }
+    } catch (err) {
+      if (err.message === 'User cancelled') {
+        showToast(JULES_MESSAGES.cancelled(totalSuccessful, totalItems), 'warn');
+        statusBar.clear();
+        await loadQueuePage();
+        return;
+      }
+      handleError(err, { source: 'runSelectedQueueItems' }, { category: ErrorCategory.UNEXPECTED });
+      statusBar.clearProgress();
+      statusBar.clearAction();
+      await loadQueuePage();
+      return;
+    }
+  }
+
+  if (totalSkipped > 0 && totalSuccessful === 0) {
+    showToast(JULES_MESSAGES.cancelled(0, totalItems), 'warn');
+  } else if (totalSkipped > 0) {
+    showToast(JULES_MESSAGES.completedWithSkipped(totalSuccessful, totalSkipped), 'success');
+  } else {
+    showToast(JULES_MESSAGES.COMPLETED_RUNNING, 'success');
+  }
+  statusBar.clear();
+  statusBar.clearAction();
+  await loadQueuePage();
+}
